Changes to GUI structure, issues atm
